Use PHP namespaces
WordPress development requires you to never conflict with a name in any plugin, which
traditionally required you to prefix every name with a unique prefix. Since the introduction
of namespaces in PHP 5.3, we have an alternative. To protect ourselves from any other plugin
conflicting with our plugin, this commit converts the project to use namespaces and PSR-4
autoloading which removes all the `require_once` lines from the project.